requirements.txt
fastapi>=0.109.0
packaging>=23.2
pydantic>=2.5.2
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
typing-extensions>=4.8.0
uvicorn>=0.24.0
aiofiles>=23.1.0
aiohttp>=3.11.11,<4.0.0
asyncio>=3.4.3
httpx>=0.25.0
python-multipart>=0.0.6
requests>=2.32.3
autogen-agentchat>=0.4.7,<0.5.0
autogen-core>=0.4.7,<0.5.0
autogen-ext[docker]>=0.4.7,<0.5.0
autogen-ext[langchain]>=0.4.7,<0.5.0
autogen-ext[magentic-one]>=0.4.7,<0.5.0
autogen-ext[openai,azure]>=0.4.7,<0.5.0
autogen-ext[web-surfer]>=0.4.7,<0.5.0
autogen-ext[diskcache]>=0.4.7,<0.5.0
composio-core>=0.6.11.post1
deepseek>=0.1.0
magentic-one-cli>=0.2.2
ollama>=0.1.5
azure-ai-inference>=1.0.0b6
azure-ai-ml>=1.12.1
azure-ai-projects>=1.0.0b4
azure-core>=1.32.0
azure-cosmos>=4.9.0
azure-identity>=1.19.0
azure-keyvault-secrets>=4.9.0
azure-search-documents==11.6.0b4
azure-storage-blob>=12.24.0
jsonlines>=4.0.0
matplotlib>=3.10.0
numpy>=1.26,<2.0
pandas>=2.2.3
pillow>=11.0.0
plotly>=5.16.0
plotly-express>=0.4.1
pytz>=2024.2
requests-html>=0.10.0
seaborn>=0.13.0
yfinance>=0.2.51
beautifulsoup4>=4.12.0
bing>=0.31
chainlit>=2.2.0
docker>=7.1.0
markitdown>=0.0.1a3
opentelemetry-instrumentation-fastapi>=0.44b0
playwright>=1.20.0  # Specify version for compatibility
sqlalchemy[asyncio]>=2.0.28
psycopg2-binary>=2.9.9
supabase>=2.3.1
tenacity>=9.0.0
aioconsole>=0.8.1
axios>=0.4.0
authlib>=1.4.0
black>=24.10.0
codecov>=2.1.13
gunicorn>=23.0.0
ipykernel>=6.29.5
psutil>=5.9.6
puremagic>=1.28
pytest>=8.3.4
pytest-asyncio>=0.25.2
pytest-cov>=6.0.0
redis>=5.2.1
rich>=13.9.4
ruff==0.4.8
scikit-learn>=1.6.1
starlette>=0.41.3
semantic-kernel>=1.19.0
tiktoken>=0.8.0
typer>=0.15.1
types-setuptools>=75.8.0.20250110
pyyaml>=6.0.2
setuptools>=75.8.0
macholib>=1.16.3

